+++ CHANGELOG.md
@@ -24,6 +24,7 @@
   - Update AppBidding API:
     - Replace `Criteo#setBidsForAdUnit(Object, AdUnit)` method by `Criteo#enrichAdObjectWithBid(Object, Bid)`
     - Use `Criteo#loadBid(AdUnit, BidResponseListener)` method to provide `Bid` object
+  - Make `CriteoNativeAdListener` an interface instead of an abstract class
 
 - Features
   - CriteoBannerView is now a custom view that can be included directly in a layout file.